backend/requirements.txt
fastapi
uvicorn
spotipy
python-dotenv
requests
pydantic 
yt-dlp 
librosa 
numpy
lyricsgenius
transformers
torch
